move modal to a separate component
